refactor: setup for admin api
