+++ routes/api.routes.js
@@ -1,17 +1,17 @@
-import { registrationHandler } from "../services/registration.js";
-import { registrationSchema } from "../services/schemas/registrationSchema.js";
+import { registrationHandler } from "../services/auth/registration.js";
+import { registrationSchema } from "../services/auth/schemas/registrationSchema.js";
 
-import { loginSchema } from "../services/schemas/loginSchema.js";
-import { loginHandler } from "../services/login.js";
+import { loginSchema } from "../services/auth/schemas/loginSchema.js";
+import { loginHandler } from "../services/auth/login.js";
 
-import { userProfileSchema } from "../services/schemas/userProfileSchema.js";
-import { userProfileHandler } from "../services/profile.js";
+import { userProfileSchema } from "../services/auth/schemas/userProfileSchema.js";
+import { userProfileHandler } from "../services/auth/profile.js";
 
-import { profileRecoverySchema } from "../services/schemas/profileRecoverySchema.js";
-import { profileRecoveryHandler } from "../services/recovery.js";
+import { profileRecoverySchema } from "../services/auth/schemas/profileRecoverySchema.js";
+import { profileRecoveryHandler } from "../services/auth/recovery.js";
 
-import { tokenRefreshHandler } from "../services/refresh.js";
-import { refreshSchema } from "../services/schemas/refreshSchema.js";
+import { tokenRefreshHandler } from "../services/auth/refresh.js";
+import { refreshSchema } from "../services/auth/schemas/refreshSchema.js";
 
 import { registrationOptionsHandler } from "../services/webAuthn/registrationOptions.js";
 import { registrationVerificationHandler } from "../services/webAuthn/registrationVerification.js";
@@ -21,7 +21,8 @@ import { loginVerificationHandler } from "../services/webAuthn/loginVerification
 import { authenticateRequest } from "../utilities/authenticate.js";
 
 const serverRoutes = async function (fastify, options) {
-  //auth API routes
+  
+  //authentication API routes
   fastify.post("/register", registrationSchema, registrationHandler);
   fastify.post("/login", loginSchema, loginHandler);
   fastify.post(
@@ -32,11 +33,14 @@ const serverRoutes = async function (fastify, options) {
   fastify.post("/recovery", profileRecoverySchema, profileRecoveryHandler);
   fastify.post("/refresh", refreshSchema, tokenRefreshHandler);
 
-  //webAuthn routes
+  //webAuthn API routes
   fastify.get("/registration-options", registrationOptionsHandler);
   fastify.post("/registration-verification", registrationVerificationHandler);
   fastify.get("/login-options", loginOptionsHandler);
   fastify.post("/login-verification", loginVerificationHandler);
+
+  //admin API routes
+
 };
 
 export default serverRoutes;
